update(web): update `task-add-label` for compatibility with tags
